Transformations API tests optimization
