+++ Shared.Tests/CloudinaryTest.cs
@@ -2117,359 +2117,6 @@ namespace CloudinaryDotNet.Test
         }
 
         [Test]
-        public void TestListTransformations()
-        {
-            // should allow listing transformations
-
-            ImageUploadParams uploadParams = new ImageUploadParams()
-            {
-                File = new FileDescription(m_testImagePath),
-                EagerTransforms = new List<Transformation>() { m_simpleTransformation },
-                Tags = m_apiTag
-            };
-
-            m_cloudinary.Upload(uploadParams);
-
-            ListTransformsResult result = m_cloudinary.ListTransformations();
-
-            Assert.IsNotNull(result);
-            Assert.IsNotNull(result.Transformations);
-            TransformDesc td = result.Transformations.Where(t => t.Name == m_simpleTransformationAsString).First();
-            Assert.IsTrue(td.Used);
-        }
-
-        [Test]
-        public void TestGetTransform()
-        {
-            // should allow getting transformation data.
-
-            var uploadParams = new ImageUploadParams()
-            {
-                File = new FileDescription(m_testImagePath),
-                EagerTransforms = new List<Transformation>() { m_updateTransformation },
-                Tags = m_apiTag
-            };
-
-            m_cloudinary.Upload(uploadParams);
-
-            var result = m_cloudinary.GetTransform(m_updateTransformationAsString);
-
-            Assert.IsNotNull(result);
-            Assert.AreEqual(m_updateTransformation.ToString(), new Transformation(result.Info[0]).ToString());
-        }
-
-        [Test]
-        public void TestUpdateTransformStrict()
-        {
-            // should allow updating transformation allowed_for_strict
-
-            ImageUploadParams uploadParams = new ImageUploadParams()
-            {
-                File = new FileDescription(m_testImagePath),
-                EagerTransforms = new List<Transformation>() { m_simpleTransformation },
-                Tags = m_apiTag
-            };
-
-            m_cloudinary.Upload(uploadParams);
-
-            UpdateTransformParams updateParams = new UpdateTransformParams()
-            {
-                Transformation = m_simpleTransformationAsString,
-                Strict = true
-            };
-
-            m_cloudinary.UpdateTransform(updateParams);
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(m_simpleTransformationAsString);
-
-            Assert.IsNotNull(getResult);
-            Assert.AreEqual(true, getResult.Strict);
-
-            updateParams.Strict = false;
-            m_cloudinary.UpdateTransform(updateParams);
-
-            getResult = m_cloudinary.GetTransform(m_simpleTransformationAsString);
-
-            Assert.IsNotNull(getResult);
-            Assert.AreEqual(false, getResult.Strict);
-        }
-
-        [Test]
-        public void TestUpdateTransformUnsafe()
-        {
-            string transformationName = GetUniqueTransformationName();
-            // should allow unsafe update of named transformation
-            m_cloudinary.CreateTransform(
-                new CreateTransformParams()
-                {
-                    Name = transformationName,
-                    Transform = m_simpleTransformation
-                });
-
-            var updateParams = new UpdateTransformParams()
-            {
-                Transformation = transformationName,
-                UnsafeTransform = m_updateTransformation
-            };
-
-            m_cloudinary.UpdateTransform(updateParams);
-
-            var getResult = m_cloudinary.GetTransform(transformationName);
-
-            Assert.IsNotNull(getResult.Info);
-            Assert.AreEqual(updateParams.UnsafeTransform.Generate(), new Transformation(getResult.Info).Generate());
-        }
-
-        [Test]
-        public void TestUpdateTransformWithImplicitExtension()
-        {
-            // should allow updating a transformation string contains a format extension (e.g., w_150,h_150/jpg) or an empty extnetion (e.g., w_150,h_150/)
-
-            string transformationNameWithExt = m_updateTransformationWithExtAsString;
-            string transformationNameWithEmptyExt = m_updateTransformationWithEmptyExtAsString;
-            
-            var create = new CreateTransformParams() { Transform = m_updateTransformation };
-
-            create.Name = transformationNameWithExt;
-            create.Format = FILE_FORMAT_JPG;
-
-            m_cloudinary.CreateTransform(create);
-
-            create.Name = transformationNameWithEmptyExt;
-            create.Format = String.Empty;
-
-            m_cloudinary.CreateTransform(create);
-
-            var updateParams = new UpdateTransformParams() { Strict = true };
-
-            updateParams.Transformation = transformationNameWithExt;
-            updateParams.Format = FILE_FORMAT_JPG;
-
-            m_cloudinary.UpdateTransform(updateParams);
-
-            updateParams.Transformation = transformationNameWithEmptyExt;
-            updateParams.Format = String.Empty;
-
-            m_cloudinary.UpdateTransform(updateParams);
-
-            var getResultWithExt = m_cloudinary.GetTransform(transformationNameWithExt);
-
-            Assert.IsNotNull(getResultWithExt.Info);
-            Assert.AreEqual(true, getResultWithExt.Strict);
-
-            var getResultWithEmptyExt = m_cloudinary.GetTransform(transformationNameWithEmptyExt);
-
-            Assert.IsNotNull(getResultWithEmptyExt.Info);
-            Assert.AreEqual(true, getResultWithEmptyExt.Strict);
-        }
-
-        [Test]
-        public void TestCreateTransform()
-        {
-            // should allow creating named transformation
-
-            CreateTransformParams create = new CreateTransformParams()
-            {
-                Name = GetUniqueTransformationName(),
-                Transform = m_simpleTransformation
-            };
-
-            m_cloudinary.CreateTransform(create);
-
-            GetTransformParams get = new GetTransformParams()
-            {
-                Transformation = create.Name
-            };
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(get);
-
-            Assert.IsNotNull(getResult);
-            Assert.AreEqual(true, getResult.Strict);
-            Assert.AreEqual(false, getResult.Used);
-            Assert.AreEqual(1, getResult.Info.Length);
-            Assert.AreEqual(m_simpleTransformation.Generate(), new Transformation(getResult.Info[0]).Generate());
-        }
-
-        [Test]
-        public void TestCreateTransformWithExt()
-        {
-            // should allow creating a transformation containing an extension suffix (e.g., w_150,h_150/jpg)
-
-            CreateTransformParams create = new CreateTransformParams()
-            {
-                Name = m_updateTransformationWithExtAsString,
-                Transform = m_updateTransformation,
-                Format = FILE_FORMAT_JPG
-            };
-
-            m_cloudinary.CreateTransform(create);
-
-            GetTransformParams get = new GetTransformParams()
-            {
-                Transformation = create.Transform.Generate(),
-                Format = FILE_FORMAT_JPG
-            };
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(get);
-
-            Assert.IsNotNull(getResult);
-            Assert.AreEqual(true, getResult.Strict);
-            Assert.AreEqual(false, getResult.Used);
-            Assert.AreEqual(1, getResult.Info.Length);
-
-            string formatStr = null;
-
-            if (getResult.Info[0].TryGetValue("format", out var format))
-            {
-                formatStr = format.ToString();
-            }
-
-            Assert.AreEqual(FILE_FORMAT_JPG, formatStr);
-            Assert.AreEqual(m_updateTransformationAsString, new Transformation(getResult.Info[0]).Generate().ToString());
-        }
-
-        [Test]
-        public void TestCreateTransformWithEmptyExt()
-        {
-            // should allow creating a transformation ending with a slash (e.g., w_150,h_150/)
-
-            CreateTransformParams create = new CreateTransformParams()
-            {
-                Name = m_updateTransformationWithEmptyExtAsString,
-                Transform = m_updateTransformation,
-                Format = String.Empty
-            };
-
-            m_cloudinary.CreateTransform(create);
-
-            GetTransformParams get = new GetTransformParams()
-            {
-                Transformation = create.Transform.Generate(),
-                Format = String.Empty
-            };
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(get);
-
-            Assert.IsNotNull(getResult);
-            Assert.AreEqual(true, getResult.Strict);
-            Assert.AreEqual(false, getResult.Used);
-            Assert.AreEqual(1, getResult.Info.Length);
-
-            string extensionStr = null;
-
-            if (getResult.Info[0].TryGetValue("extension", out var extension))
-            {
-                extensionStr = extension.ToString();
-            }
-
-            Assert.AreEqual("none", extensionStr);
-            Assert.AreEqual(m_updateTransformationAsString, new Transformation(getResult.Info[0]).Generate().ToString());
-        }
-
-        [Test]
-        public void TestDeleteTransform()
-        {
-            // should allow deleting named transformation
-            string transformationName = GetUniqueTransformationName();
-
-            CreateTransformParams create = new CreateTransformParams()
-            {
-                Name = transformationName,
-                Transform = m_simpleTransformation
-            };
-
-            TransformResult createResult = m_cloudinary.CreateTransform(create);
-
-            Assert.AreEqual("created", createResult.Message);
-
-            var res = m_cloudinary.DeleteTransform(transformationName);
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(
-                new GetTransformParams() { Transformation = transformationName });
-
-            Assert.AreEqual(HttpStatusCode.NotFound, getResult.StatusCode);
-        }
-
-        [Test]
-        public void TestDeleteTransformWithExt()
-        {
-            // should allow deleting a transformation containing an extension suffix (e.g., w_150,h_150/jpg)
-            string transformationName = m_simpleTransformationWithExtAsString;
-
-            CreateTransformParams create = new CreateTransformParams()
-            {
-                Name = transformationName,
-                Transform = m_simpleTransformation,
-                Format = FILE_FORMAT_JPG
-            };
-
-            TransformResult createResult = m_cloudinary.CreateTransform(create);
-
-            Assert.AreEqual("created", createResult.Message);
-
-            m_cloudinary.DeleteTransform(transformationName);
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(
-                new GetTransformParams() { Transformation = transformationName });
-
-            Assert.AreEqual(HttpStatusCode.NotFound, getResult.StatusCode);
-        }
-
-        [Test]
-        public void TestDeleteTransformWithEmptyExtension()
-        {
-            // should allow deleting a transformation ending with a slash (e.g., w_150,h_150/)
-            string transformationName = m_simpleTransformationWithEmptyExtAsString;
-
-            CreateTransformParams create = new CreateTransformParams()
-            {
-                Name = transformationName,
-                Transform = m_simpleTransformation,
-                Format = String.Empty
-            };
-
-            TransformResult createResult = m_cloudinary.CreateTransform(create);
-
-            Assert.AreEqual("created", createResult.Message);
-
-            m_cloudinary.DeleteTransform(transformationName);
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(
-                new GetTransformParams() { Transformation = transformationName });
-
-            Assert.AreEqual(HttpStatusCode.NotFound, getResult.StatusCode);
-        }
-
-        [Test]
-        public void TestDeleteTransformImplicit()
-        {
-            // should allow deleting implicit transformation
-            ImageUploadParams uploadParams = new ImageUploadParams()
-            {
-                File = new FileDescription(m_testImagePath),
-                EagerTransforms = new List<Transformation>() { m_implicitTransformation },
-                Tags = m_apiTag
-            };
-
-            m_cloudinary.Upload(uploadParams);
-
-            GetTransformParams getParams = new GetTransformParams()
-            {
-                Transformation = m_implicitTransformation.ToString()
-            };
-
-            GetTransformResult getResult = m_cloudinary.GetTransform(getParams);
-
-            Assert.AreEqual(HttpStatusCode.OK, getResult.StatusCode);
-
-            m_cloudinary.DeleteTransform(m_implicitTransformation.ToString());
-
-            getResult = m_cloudinary.GetTransform(getParams);
-
-            Assert.AreEqual(HttpStatusCode.NotFound, getResult.StatusCode);
-        }
-
-        [Test]
         public void TestUploadHeaders()
         {
             ImageUploadParams uploadParams = new ImageUploadParams()
@@ -3651,20 +3298,6 @@ namespace CloudinaryDotNet.Test
                 Type = STORAGE_TYPE_UPLOAD,
                 ResourceType = ResourceType.Image
             });
-        }
-
-        [Test]
-        public void TestArchiveParamsCheck()
-        {
-            Assert.Throws<ArgumentException>(new ArchiveParams().Check, "Should require atleast on option specified: PublicIds, Tags or Prefixes");
-        }
-        [Test]
-        public void TestCreateTransformParamsCheck()
-        {
-            var p = new CreateTransformParams();
-            Assert.Throws<ArgumentException>(p.Check, "Should require Name");
-            p.Name = "some_name";
-            Assert.Throws<ArgumentException>(p.Check, "Should require Transformation");
         }
 
         [Test]
